matcher: collapse the hand-written matchers into the framework (single source)
With both back-ends now derived from the match-plan framework, the two
hand-written per-construct dispatches are redundant and are removed, leaving the
framework as the single source of truth for find-matching:

- delete CompiledMatchProgram's term-level dispatch (compile / compileStep /
  compileCore / compile{ElementaryUpdate,ParametricFunction,Modality}); the heads
  already carry that logic. Its reused Java-program helpers (compiledProgramMatcher,
  compileProgram, delegateToMatch, compileActiveStatements) move to a small
  JavaProgramCompiler used by the program hook.
- delete SyntaxElementMatchProgramGenerator's createProgram dispatch; only the
  program-instruction conversion helpers (buildProgramInstruction & co), which the
  hook reuses, remain.
- migrate term labels into the framework (JavaMatchPlanBuilder.LabelPlan, reusing
  the matchTermLabelSV instruction) so buildPlan is total; the facades no longer
  fall back -- an unsupported pattern raises a clear error naming the missing head
  (no current taclet hits this; the whole standard base is covered).
- retarget CompiledMatchProgramTest to the framework facade.

Net: ~390 fewer lines of production matcher code, no duplicated dispatch. The
interpreter/compiled engines and the program helpers are unchanged.
